/article/id URL param must be a `path`
When a URL parameter is defined as `string` and it contains `/`, then
the web API is confused and it considers it as multiple params.

You need to define it as `id:path`

Ref: https://fastapi.tiangolo.com/tutorial/path-params/#path-convertor

Also note that when a URL param is a `path`, it must be the last one in
the URL for obvious reasons.